+++ src/platform/assets/components/Media3DTop.vue
@@ -10,17 +10,8 @@
       v-else
       class="flex size-full flex-col items-center justify-center gap-2 bg-modal-card-placeholder-background transition-transform duration-300 group-hover:scale-105 group-data-[selected=true]:scale-105"
     >
-      <i
-        :class="
-          cn(
-            'text-3xl text-muted-foreground',
-            loading
-              ? 'icon-[lucide--loader-circle] animate-spin'
-              : 'icon-[lucide--box]'
-          )
-        "
-      />
-      <span v-if="!loading" class="text-sm text-base-foreground">
+      <i class="icon-[lucide--box] text-3xl text-muted-foreground" />
+      <span class="text-sm text-base-foreground">
         {{ $t('assetBrowser.media.threeDModelPlaceholder') }}
       </span>
     </div>
@@ -31,22 +22,15 @@
 import { useIntersectionObserver } from '@vueuse/core'
 import { onBeforeUnmount, ref, watch } from 'vue'
 
-import { cn } from '@/utils/tailwindUtil'
-
-import { renderThumbnail } from '@/extensions/core/load3d/ThumbnailRenderer'
 import { api } from '@/scripts/api'
 
 import type { AssetMeta } from '../schemas/mediaAssetSchema'
-import {
-  findServerPreviewUrl,
-  persistThumbnail
-} from '../utils/assetPreviewUtil'
+import { findServerPreviewUrl } from '../utils/assetPreviewUtil'
 
 const { asset } = defineProps<{ asset: AssetMeta }>()
 
 const containerRef = ref<HTMLElement>()
 const thumbnailSrc = ref<string | null>(null)
-const loading = ref(false)
 const hasAttempted = ref(false)
 
 useIntersectionObserver(containerRef, ([entry]) => {
@@ -68,24 +52,7 @@ async function loadThumbnail() {
     const serverPreviewUrl = await findServerPreviewUrl(asset.name)
     if (serverPreviewUrl) {
       thumbnailSrc.value = serverPreviewUrl
-      return
     }
-  }
-
-  loading.value = true
-  try {
-    const blob = await renderThumbnail(asset.src)
-    if (!blob) return
-
-    thumbnailSrc.value = URL.createObjectURL(blob)
-
-    if (asset.id && api.getServerFeature('assets', false)) {
-      void persistThumbnail(asset.name, blob)
-    }
-  } catch {
-    thumbnailSrc.value = null
-  } finally {
-    loading.value = false
   }
 }
 
@@ -102,7 +69,6 @@ watch(
     if (hasAttempted.value) {
       hasAttempted.value = false
       revokeThumbnail()
-      loading.value = false
     }
   }
 )
